+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Release
 
+## 0.0.16
+
 - [sdk] Fix cancellation.
 - [sdk] Change method name from `cancelExpired` to `cancel`.
 - [sdk, router] Add `network` param for abstracting messaging configuration.